// src/managed/ingress-env.d.ts
/* eslint-disable */
// Generated by Wrangler by running `wrangler types src/managed/ingress-env.d.ts -c managed/ingress.json --env-interface ManagedIngressEnv --include-runtime false` (hash: 8037bdd5e8286594abd268bae691d292)
interface __BaseEnv_ManagedIngressEnv {
  MANAGED_CONFIG: KVNamespace;
  MANAGED_STAGE: 'local-scaffold';
  MANAGED_AUTHORIZATION: DurableObjectNamespace<import('./ingress').ManagedAuthorization>;
  MANAGED_DELIVERY: Fetcher /* bugdrop-managed-local-delivery */;
}
declare namespace Cloudflare {
  interface GlobalProps {
    mainModule: typeof import('./ingress');
    durableNamespaces: 'ManagedAuthorization';
  }
  interface Env extends __BaseEnv_ManagedIngressEnv {}
}
interface ManagedIngressEnv extends __BaseEnv_ManagedIngressEnv {}
type StringifyValues<EnvType extends Record<string, unknown>> = {
  [Binding in keyof EnvType]: EnvType[Binding] extends string ? EnvType[Binding] : string;
};
declare namespace NodeJS {
  interface ProcessEnv extends StringifyValues<Pick<Cloudflare.Env, 'MANAGED_STAGE'>> {}
}

export { ManagedIngressEnv };