// src/managed/delivery-env.d.ts
/* eslint-disable */
// Generated by Wrangler by running `wrangler types src/managed/delivery-env.d.ts -c managed/delivery.json --env-interface ManagedDeliveryEnv --include-runtime false` (hash: c974573a3f671e048b15a25ef456f15a)
interface __BaseEnv_ManagedDeliveryEnv {
  MANAGED_OUTCOMES: Queue;
  MANAGED_STAGE: 'local-scaffold';
  MANAGED_RECEIPTS: DurableObjectNamespace<import('./delivery').ManagedReceipts>;
}
declare namespace Cloudflare {
  interface GlobalProps {
    mainModule: typeof import('./delivery');
    durableNamespaces: 'ManagedReceipts';
  }
  interface Env extends __BaseEnv_ManagedDeliveryEnv {}
}
interface ManagedDeliveryEnv extends __BaseEnv_ManagedDeliveryEnv {}
type StringifyValues<EnvType extends Record<string, unknown>> = {
  [Binding in keyof EnvType]: EnvType[Binding] extends string ? EnvType[Binding] : string;
};
declare namespace NodeJS {
  interface ProcessEnv extends StringifyValues<Pick<Cloudflare.Env, 'MANAGED_STAGE'>> {}
}

export { ManagedDeliveryEnv };